R238: fix AniSkip review blockers and migration safety

+++ app/src/main/java/eu/kanade/tachiyomi/ui/player/utils/AniSkipCacheEntry.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 internal data class AniSkipCacheEntry(
     val malId: Long,
-    val episodeNumber: Int,
+    val episodeNumber: Double,
     val roundedEpisodeLength: Long,
     val createdAt: Long,
     val timestamps: List<TimeStamp>,
